Added Party Confirmation to Party Assignment.
Just like it says, I added the basic party assignment based on the y value of the character panels. Will still need to implement the pfps for the party members, but I will wait for Cici to do that.

Also need to implement bparties in the Party script. It's the backup party that will be saved to ensure when the menu is navigated to again, the parties easily collapse back into a single party for ease of use. Should probably implement the "reassembleparty" method in the Party script to achieve this.