Moved back to reimplementing the Combat system
Fixed the Potato Yeet problem by removing Mirror and then individually removing the components it created.

Likewise, I also updated the enemies to now all have health bars and stats with them (even if they are all 0 except for HP and SP). Likewise, for testing purposes, I copied the Party GameObject from the Party Assignment scene and temporarily placed it in the Combat scene.

TODO: Fix the Combat UI because it is not set for 1920x1080 atm. Will also need to anchor all the UI elements to potentially the bottom left corner or so.