2 steps forward, 9 steps back
Added a better UI to the game for menu navigation.

Added skill pictures to the skill buttons

Upon trying to readjust the party a bunch, it became prevalent that an extra dictionary for bparties won't work because the reference value for the lists are references, not exact values. So they share it. Will need to perform a deep copy or a clone of the original parties and all the lists inside of it, then I can put them into bparties.

Also added upon party confirmation, the menu opener returning to the hierarchy so that the menu can be reopened again.