Various changes
Added ambushed song.

Updated the Party to directly reference the Players' GameObjects. Will need to reassign the Party's values in the PartyAssignment Scene.

Added the Looper to now loop songs that have an intro.

Current issue: Although technically the values being subtracted and whanot are indeed directly related to the player value in the Party, it's not being reflected in the Inspector. Will need to fix this later.