Gun pos and scale added a while ago. Fixed "bouncing" issue and a "workaround"
Each gun model was set correctly into Tao's hands with the correct scaling and put into a text file. Will add into the code later on for the battle sequences.

Also fixed the issue with the player controller "bouncing" by manually setting y component to 0 of the move vector. This caused an issue with speed, so the top rig of the cinemachine was lowered to make the slow-down less noticeable.