Added movement for Player Combat Controller
Added a goTo() method with a list of coordinates that the player will have to go to in order to attack the enemy.

Likewise, I also added a returnToSpawn method which will move the players back to their original battle area.

TODO: Make the player face the enemy they are running to, and back to their area, then face the battlefield again upon reaching their area.