Added the "spell casting" track for P1 and misc changes to Battle System
Created 2 more waypoints on the track, namely 2 and 3 for player 1's summoning animation "orbit" which may need refinement.

Made various changes to the Battle System code to accommodate for a few bugs found, as well as making the code more.... optimized. Also added code to the spellcasting method instructing the camera to travel between waypoints 2 and 3. Will test at a later date.

Added a lot of TODOs as well including where some animations may need to be played. Will also need to add more pauses to allow animations to play further. There are still a LOT of bugs to fix in that logic, but they will be fixed in due time.

PlayerCombatController got reformatted by Visual Studio, no changes there.