Added Most Target Logic.
A target game object, holding the health bar too, was added to all enemies, and a simple target game Object was likewise added to players.

The TargetManager script would reveal the target, make it face the camera's position, and move the target a bit closer to the camera so that strange clipping is limited/nonexistant.

Still need to figure out how to approach multiple targetting efficiency, cursor read activation, etc. I figure scripting a raycast after moving into the "targetting" phase would be the best option, but will need to be tested.

Altered Unit to now no longer use the slider, but the filled image of actual health bar. Will still need to animate the fill amount decreasing over time as well as facing the camera when the enemy is attacked. The players *can* have it too, but it's better to leave the animation to the corner head pieces. May use this in the TakeDamage method in Unit.

Also fumbled around a bit with tags and layers trying to figure out how to see target the best. Possibly will need to go back in and readjust those things to ensure the most organized space.

May also need to reangle cameras depending on how the player heads look in comparison to the rest of the combat scene.... but that may be something to worry about for later.