Added a bit more optimization to the BattleSystem and Removed DefaultBattleManager
Added a better check for enemy spawns to make it a bit... cleaner to look at. Added better checks to make sure the players move to the correct area based on the certain checks.

Removed DefaultBattleManager as it was not really contributing to anything and at this point, the Battle System is better equipped to be converted to NetworkBehaviour.

Altered the enemy prefabs to correctly use the healthbar Image.