Fixed Targetting: Moving to Attacking
Target no longer slowly gets higher from aiming at different people.
Target will be reset when to inactive when the move is not currently hovering over it.

Camera will now move to an elevated position for targetting. Possibly will need to make the travel to it a bit quicker.
Pressing escape while targetting will cancel the targetting mode and return the camera back down to the correct player default camera.
CinemachineCombatHandler cameras were all removed, replaced by the state driven camera and the initial dolly cart camera.

Next steps:
Implement the player either running back to their designated spot while the next turn plays, and resetting their rotation, or teleporting them back if it's a 1 more. Will also need to make sure the player is looking at the enemy while attacking them, but will need to implement attacking animation while doing so.
Go back and reorganize the code so it's not so confusing and spaghetti-like, as the CCH script's lookTarget function is sometimes pretty confusing.
IMPLEMENT BATON PASSING FUNCTION AT A LATER TIME!!!
After the physical attack is finished, the shooting part will be done, and will use a very similar process to the physical attack, but will be third person, have a crosshair, and will provide an option to shoot at teammates, triggering an instant miss and an angry voiceline (will be added later).

Possible bugs (untested):
Test to make sure the player even gets to the enemy before the damage is calculated.
Test to make sure the player rotates to where they are going. It could possibly not be rotating while they approach enemies and when they return to their spot.

Known bugs:
If the player hovers over an enemy before the camera reaches it's desired area, the target will appear strange. This could potentially be fixed with a Lerp to where the camera will eventually be. Other avenues are possible.