Fixed Player 1 not moving issue
Battle System:
Changed the getPlayerObject() method to return the GameObject directly
Changed tghe getPlayer() method to getPlayerInParty() which will be used for health, SP, ailment, etc. statuses. May need to go back through the script itself and make sure that it calls the correct methods since it's easy to confuse the instantiated cosmetic object and the actual player within the party.
Started adding *some* of the logic for shooting mechanics, but will be worked on later, as physical attacks come first.

Unit:
Added a resistanceChecker() method to return values, making a weakness check a bit easier to compute and look at.

Player Combat Controller:
Adjusted bits of the movement script and stuff to make it function a bit better. Including changing the check from the move.magnitude being 0.1 and now being 0.2.
Will potentially need to adjust the speed of the characters running in to be a default 6, but with a starting velocity of like 3 for running in. This should be written alongside the Battle System script since it has like 4 instances where the speed is set to 6.

Player:
Added unconscious = true for when the HP is <=0 after taking damage.

Cinemachine:
Should probably add some sort of script handler to as soon as the player reaches the enemy, will focus on the enemy instead of the player, potentially adding another VCam in it's place.
Also VCams around for Players 2-4 for default reference.

Known Bugs:
Select.Play() still has an issue and will need to be fixed at a later date
Party.Start() still needs to be removed when the time comes