Characters Rotate Now
Players can now face their enemy, the enemy faces them, and they attack. Afterwards, they turn around and run back to their spawn and accurately turn back around to face the battlefield.

All builds are given the name PersonaStoryCognitionHackers now.

Note:
Still need to adjust the position the player runs to in order to attack. I believe I will just change all that logic for the player to just run at the enemy with an offset of magnitude and proceed to smack the crap out of the enemy.

Will do this by following the if(isAttacking) statement up with an additional condition for the player to make sure they aren't *quite* to the enemy before they start attacking. This will also make sure that the players don't accidentally just sidestep the enemy.

Upon the circle being someone else's turn, the UI should show something different than the actual attack menu and stuff.

WARNING:
Will need to alter the logic a bit for the *actual* players so that they can actually interact in the battle. Will need to disable the PC script during battle, and reenable outside of battle. Likewise, I need to enable PCC during battle, and disable afterwards to prevent strangeness.