Changed "Player" Script to "Persona" Script
Changed the Player script to the Persona script in order to separate the combat data from the player data itself, as both aren't needed to be read at the same time. Will need to make slight adjustements to these scripts, like switching the "money" value over to the Player and directly adding money gained per battle in the Player script, as well as the exp and so on in the Persona script.

NOTE:
Will need to update the UpdatePlayerData() method to also accept the Player playerFile and update the values inside of it in the Realtime Database for the FirebaseManager script. Will add this as an "issue" for future use