Added *some* gun animations and logic
Created Sciptable Objects for the melee weapons and guns for more functionality when it comes to actual damage and stuff. Will need to add the weapons and values to a Scriptable Object File.

Added Gun logic, like bullet counting, and started the damage logic for shooting enemies and players, as well as playing the gunshot sound on an AudioSource attached to the player. Realized very quickly I will need to add the missing and doging chances at this point in order to effectively make a "dodge" mechanic for the player. The "angry voice" part can be coded in later once I have actual voice acting, and can be used with an extra AudioSource in the scene probably just called "intercom voice" or something.

Made the animator for the PCC public so the Battle System can make the player "shoot" or pull out their guns. Can also be super helpful for letting a player go ahead and attack an enemy in the Battle System instead of needing to reference, and code in the animation itself into the Persona script.

Will need to add armor as a Scriptable Object soon to help accommodate for evasion and defense as well, better sooner than later.

Will also need to put references on player's hands so I can swap between guns and melee effectively. Will also need to write in the logic of switching between both as well.

NOTE:
Still need to update the player running up to attack the enemy with actually running up to the enemy instead of running to a position to avoid issues with placement.

Need to add the missing and doging chances at this point in order to effectively make a "dodge" mechanic for the player and enemy.
Need to add the chance that if the player misses the enemy in melee attack, there's a 5% chance they will go to the "isDown" state, at which the camera needs to see them miss and fall.
It will probably be best to calculate the dodge chance through referencing the numbers into the Battle System just to make things.... cleaner and simpler.