Started making the miss, crit, and evade scripts
BattleSystem:
Added miss conditions that for melee, there is a 5% chance that you fall to the "Down" position. WILL NEED TO ALSO IMPLEMENT THAT FOR ENEMIES TO FALL AS WELL SO IT'S FAIR!!!! Added a few more TODOs at the top of the script in regards to faux numbers to account for when crits are boosted and whatnot. Various other improvements (hopefully, I never tested them).
Added playerMissChecker() and critChecker() (which will need to be renamed to playerCritChecker()). Also added those checks into the TakeDamage() method for the enemies. WILL ALSO NEED TO APPLY THIS TO PLAYERS AT ANOTHER POINT.
Added variance to the playerDamageCalculator() method, and will need to do the same after moving on to the magic casting side of it.

Unit:
Added a crit checker into the TakeDamage() method to help make crit damage more... universal?

GunSO && WeaponSO:
Made critBoost public. Made chance into an int instead of a float.

WARNING:
This commit has not been tested like.... at all! Proceed with caution!!!

NOTE:
Armor will need to be made soon so that it can be taken into account for pretty much everything at this point. Enemy methods will need heavy remastering to be similar to the player methods. Remember to also make sure the gun attacks are relatively close to totalDamage/magazine so that each bullet, when combined, will be somewhere around the point desired.