Added Blocking Animation
As stated in title, added Tao's potential blocking animation, however it needs slight adjustments, as it is currently a little weird looking.

Added Extra placeholders into the combat and party placeholders in order to not affect the prefab info.

Added text file with most spell name(s) and their progression in power.