Added Skill functionality
Revamped skill SOs.

Made a skill reader based on the party member on the persona panel. Also updated BattleSystem to make requests to the skill reader for skill cleanup and population, again depending on the character.

Moved all files to the proper place

TODO: Add description tabs for skills. Maybe resizing the circle? Keep in mind, Nex will provide something for this too.