Started working on Nirvana VR Hub
Started modelling the Nirvana VR Hub after the Fort Worth Stock Yards. May change it in the future, but will need to design a logo for the armor and weapons shop.

Added a few Unity Assets from the store, but may have also compromised a shader, or the pipeline itself from trying to implement the 1GB MODULAR BUILDING package.

May need to go through and prune out demo scenes and the like to ensure unneeded files aren't in the game because I forgot.