Much progress.... I hope
Updated Tao's shader to hopefully be more realistic. May need to check lighting conditions to MAKE SURE shadows don't glitch out on characters' faces.

Revamped some of the Main Menu stuff. Added an intro (may have to stoop into PlayerPrefs to keep this from playing every time since it can be annoying maybe, or add a listener for ANY input, to skip it.) Need to populate the main menu with more stuff, and a listener to adjust it to be more.... interesting.

Added more sounds for the intro, renaming the previous BG music, as lobby waiting music, and adding the tru background music in.