Slight adjustments for the future
Made some adjustments for future things. I updated the Nirvana Post Processing to be more "blue"

Created a new folder for "Digital" models. Adjusting the combat script to enable and disable which controller is needed for each situation, whether that is a normal player controller, or a combat controller. Will need to make these changes for the Real World scenes as well.

Fixed weird bug that prevented character from descending stairs without "hopping". May need to add a few more colliders to prevent the players from "stepping" up on the bench. Will definitely need to make triggers for stores in Nirvana.