Added Various Game Managers
Added the beginnings of NPC bond creation. Started with making a World Manager to handle the dates and other data, then made a Location Manager, which will need to be placed in EVERY scene, both as a backup to make sure the World Manager isn't for some reason broken, and to give info of what NPC is where. May use it for other things as well.

Created a basic NPC bond script to handle whether or not they will even spawn to begin with.