Significant changes to Intro and menu
Made a lot of changes to the intro. Only need to incorporate a sound upon pressing "Start", including the studio icon, make the options menu actually usable, adjust the main menu to be a bit more.... stylized... Need to animate the menu and buttons appearing... And finally, I need to populate the background with the characters and some NPCs.