Added more stuff for Tao and CO
Need to adjust the Persona Reader Script and add it to the Real World Counterparts, or just add the combat controller to the real world ones and don't use them.

Added more animations for Tao, including an attack that triggers combat if it hits a shadow (still needs to be coded) and used Animation Events in order to make it freeze the player with Persona Finder.

Added misc animations from PRG character animation pack so I can pick and choose some of them maybe for Tao or other characters.