Added Haruka's first and second Rapier and other changes
I also made some special fixes like fixing some of, and breaking other parts of, the combat system.

I overhauled and now the entire battlesystem is reliant on the instantiated avatars instead of the party. Likewise, the party's HP and SP is synchronized with the battle avatar's. However, not all of the other stats are and should be added to later.

I also added some extra Post Processing and neat features to having low health, like Perlin noise, Vignette pulsations, and muffled music hearing.

Battle System:
Updated everything to be "local" avatars instead of "global" party members. Also added on with a few things in regards to Post Process and combat viability.